+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * module name is now `ModelicaScriptingTools` instead of `MoST`
 * `getVariableFilter` and `getSimulationSettings` now throw `MoSTError`s if the requested model does not exist
 * use `__MoST_experiment` instead of `__ChrisS_testing` as vendor specific annotation
+* `refDir` parameter in `testmodel` and `regressionTest` now is relative to CWD instead of output dir
 
 ### Fixed